requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dotenv-python==0.0.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.3
idna==2.10
itsdangerous==2.0.0a1
Jinja2==2.11.2
MarkupSafe==2.0.0a1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
urllib3==1.26.1
Werkzeug==1.0.1
Flask-WTF==0.14.3
flask-nav==0.6
xlrd==1.2.0
psycopg2==2.8.6


